허용 앱 실행 중 다른 앱 접속 감지 및 차단

+++ app/src/main/java/com/rocket/cosmic_detox/data/model/AllowedApp.kt
@@ -7,5 +7,5 @@ import kotlinx.parcelize.Parcelize
 data class AllowedApp(
     val packageId: String = "",
     val appName: String = "",
-    val limitedTime: Int = 0,
+    val limitedTime: Long = 0,
 ) : Parcelable
